+++ src/MaaWpfGui/Views/Dialogs/TextDialogView.xaml
@@ -1,5 +1,5 @@
 <hc:Window
-    x:Class="MaaWpfGui.Views.Dialogs.TextDialogUserControl"
+    x:Class="MaaWpfGui.Views.Dialogs.TextDialogView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:controls="clr-namespace:MaaWpfGui.Styles.Controls"
