+++ src/MaaWpfGui/Views/Dialogs/TextWithTimerDialogView.xaml
@@ -1,10 +1,10 @@
 <Border
-    x:Class="MaaWpfGui.Views.UserControl.TextDialogWithTimerUserControl"
+    x:Class="MaaWpfGui.Views.Dialogs.TextWithTimerDialogView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:hc="https://handyorg.github.io/handycontrol"
-    xmlns:local="clr-namespace:MaaWpfGui.Views.UserControl"
+    xmlns:local="clr-namespace:MaaWpfGui.Views.Dialogs"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     Width="400"
     Height="300"
